refactor: 재사용을 위한 PartyItem 및 PartyList 컴포넌트화
