+++ app/src/main/java/com/mbj/doeat/ui/screen/home/detail/DetailScreen.kt
@@ -1,7 +1,6 @@
 package com.mbj.doeat.ui.screen.home.detail
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -16,9 +15,6 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.BottomSheetScaffold
@@ -46,7 +42,6 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.core.text.isDigitsOnly
@@ -64,10 +59,10 @@ import com.mbj.doeat.data.remote.model.Party
 import com.mbj.doeat.data.remote.model.SearchItem
 import com.mbj.doeat.ui.component.LoadingView
 import com.mbj.doeat.ui.component.LongRectangleButtonWithParams
+import com.mbj.doeat.ui.component.PartyList
 import com.mbj.doeat.ui.component.ToastMessage
 import com.mbj.doeat.ui.component.YesNoDialog
 import com.mbj.doeat.ui.screen.home.detail.viewmodel.DetailViewModel
-import com.mbj.doeat.ui.theme.Beige100
 import com.mbj.doeat.ui.theme.Gray200
 import com.mbj.doeat.ui.theme.Remon400
 import com.mbj.doeat.ui.theme.Yellow700
@@ -302,7 +297,7 @@ fun PartiesSection(viewModel: DetailViewModel, partyListState: List<Party>, onCl
                     NoPartiesAvailable()
                 }
             } else {
-                PartiesList(
+                PartyList(
                     viewModel = viewModel, partyListState = partyListState,
                     modifier = Modifier.weight(1f)
                 ) {
@@ -328,25 +323,6 @@ fun NoPartiesAvailable() {
         fontSize = 20.sp,
         modifier = Modifier.padding(top = 8.dp)
     )
-}
-
-@Composable
-fun PartiesList(
-    viewModel: DetailViewModel,
-    partyListState: List<Party>,
-    modifier: Modifier,
-    onClick: () -> Unit
-) {
-    LazyColumn(
-        modifier = modifier
-    ) {
-        items(
-            items = partyListState,
-            key = { party -> party.postId }
-        ) { party ->
-            PartyItem(party = party, onJoinClick = {})
-        }
-    }
 }
 
 @Composable
@@ -516,104 +492,6 @@ fun DetailBottomSheet(
             shape = RoundedCornerShape(12.dp)
         ) {
             viewModel.changeShowCreatePartyDialog(showDialog = true)
-        }
-    }
-}
-
-@Composable
-fun PartyItem(
-    party: Party,
-    onJoinClick: (Party) -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(8.dp)
-            .background(Beige100, shape = RoundedCornerShape(12.dp))
-            .border(1.dp, Color.Gray, shape = RoundedCornerShape(12.dp))
-    ) {
-        Column(
-            modifier = Modifier.padding(16.dp)
-        ) {
-            Row(
-                horizontalArrangement = Arrangement.Center
-            ) {
-                Text(
-                    text = party.restaurantName,
-                    fontWeight = FontWeight.Bold,
-                    fontSize = 18.sp,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                    color = Color.Black,
-                    modifier = Modifier
-                        .fillMaxWidth(0.6f)
-                )
-
-                Spacer(modifier = Modifier.width(8.dp))
-
-                Text(
-                    text = party.category,
-                    fontWeight = FontWeight.Bold,
-                    fontSize = 12.sp,
-                    color = Color.Gray,
-                    modifier = Modifier.align(Alignment.CenterVertically),
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis
-                )
-            }
-
-            Text(
-                text = party.restaurantLocation,
-                fontSize = 11.sp,
-                color = Color.Black,
-                modifier = Modifier.padding(top = 4.dp)
-
-            )
-
-            Text(
-                text = party.detail,
-                fontSize = 14.sp,
-                maxLines = 2,
-                overflow = TextOverflow.Ellipsis,
-                color = Color.Gray,
-                modifier = Modifier.padding(top = 8.dp)
-            )
-
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(top = 20.dp)
-                    .clickable {
-                        onJoinClick(party)
-                    },
-                horizontalArrangement = Arrangement.Center,
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                Text(
-                    text = "${party.currentNumberPeople} / ${party.recruitmentLimit}",
-                    fontSize = 22.sp,
-                    color = Color.Black,
-                    fontWeight = FontWeight.Bold,
-                    modifier = Modifier
-                )
-
-                Spacer(modifier = Modifier.width(40.dp))
-
-                Text(
-                    text = "참가하기",
-                    fontWeight = FontWeight.Bold,
-                    color = Color.Black,
-                    modifier = Modifier
-                        .background(Remon400, shape = RoundedCornerShape(4.dp))
-                        .padding(4.dp)
-                        .clickable {
-                            /**
-                             * 채팅방 참가하기 TODO
-                             */
-                            onJoinClick
-                        }
-                )
-            }
         }
     }
 }
